+++ provider/src/main/java/com/raygun/raygun4android/RaygunSettings.java
@@ -8,11 +8,14 @@ public class RaygunSettings {
     private static final int DEFAULT_MAX_REPORTS_STORED_ON_DEVICE = 64;
     private static final String DEFAULT_CRASHREPORTING_ENDPOINT = "https://api.raygun.io/entries";
     private static final String DEFAULT_RUM_ENDPOINT = "https://api.raygun.io/events";
-
     public static final String DEFAULT_FILE_EXTENSION = "raygun4";
     public static final String RUM_EVENT_SESSION_START = "session_start";
     public static final String RUM_EVENT_SESSION_END = "session_end";
     public static final String RUM_EVENT_TIMING = "mobile_event_timing";
+    // 30 minutes in milliseconds
+    // TODO Set back to 30 * 60 * 1000 when done
+    public static final int RUM_SESSION_EXPIRY = 1 * 60 * 1000;
+
 
     private static IgnoredURLs ignoredURLs = new IgnoredURLs("api.raygun.io");
     private static HashSet<String> ignoredViews = new HashSet<>();