+++ CHANGELOG.md
@@ -18,6 +18,8 @@
   - Custom endpoints are supported now
   - You can change the default of 64 reports being stored to a lower value now
   - General API cleanup and changes to Pulse -> all RUM now
+  - Removed uuid field from RaygunUserInfo
+  - You have to enable individual products now before being able to use them.
   
 - v3.0.5: Fixing timing issue from ConnectivityManager in RaygunPostService (#41), adding null checks around intent extras being null that should never be null (XRAY-1898)